refactor: extract shared tldw client helpers
