+++ apps/packages/ui/src/services/__tests__/tldw-api-client.exports.test.ts
@@ -34,5 +34,9 @@ describe("TldwApiClient module exports", () => {
     expect(module.normalizeIngestionSourceItem).toBeTypeOf("function")
     expect(module.normalizeIngestionSourceItemsListResponse).toBeTypeOf("function")
     expect(module.normalizeIngestionSourceSyncTrigger).toBeTypeOf("function")
+    expect(module.clonePresentationVisualStyleSnapshot).toBeTypeOf("function")
+    expect(module.buildPresentationVisualStyleSnapshot).toBeTypeOf("function")
+    expect(module.normalizePersonaProfile).toBeTypeOf("function")
+    expect(module.normalizePersonaExemplar).toBeTypeOf("function")
   })
 })
