+++ apps/packages/ui/src/services/tldw/__tests__/shared-normalizers.test.ts
@@ -1,0 +1,149 @@
+import { describe, expect, it } from "vitest"
+
+import {
+  normalizeIngestionSourceListResponse,
+  normalizeReadingDigestSchedule
+} from "@/services/tldw/collections-normalizers"
+import {
+  buildPresentationVisualStyleSnapshot,
+  clonePresentationVisualStyleSnapshot
+} from "@/services/tldw/presentation-style"
+import {
+  normalizePersonaExemplar,
+  normalizePersonaProfile
+} from "@/services/tldw/persona-normalizers"
+
+describe("shared tldw normalizers", () => {
+  it("normalizes reading digest schedules with string ids and boolean flags", () => {
+    expect(
+      normalizeReadingDigestSchedule({
+        id: 12,
+        enabled: 1,
+        require_online: 0,
+        format: "html"
+      })
+    ).toMatchObject({
+      id: "12",
+      enabled: true,
+      require_online: false,
+      format: "html"
+    })
+  })
+
+  it("normalizes ingestion source list payload totals and nested ids", () => {
+    expect(
+      normalizeIngestionSourceListResponse({
+        total: "2",
+        sources: [
+          {
+            id: 4,
+            user_id: "7",
+            source_type: "archive_snapshot",
+            sink_type: "notes",
+            policy: "import_only"
+          }
+        ]
+      })
+    ).toEqual({
+      total: 2,
+      sources: [
+        expect.objectContaining({
+          id: "4",
+          user_id: 7,
+          source_type: "archive_snapshot",
+          sink_type: "notes",
+          policy: "import_only"
+        })
+      ]
+    })
+  })
+
+  it("clones presentation style snapshots without reusing nested references", () => {
+    const original = buildPresentationVisualStyleSnapshot({
+      id: "style-1",
+      scope: "builtin",
+      name: "Blueprint",
+      description: "Structured deck",
+      category: "technical",
+      guide_number: 7,
+      tags: ["timeline"],
+      best_for: ["systems explanation"],
+      generation_rules: {
+        pacing: {
+          density: "tight"
+        }
+      },
+      artifact_preferences: ["comparison"],
+      appearance_defaults: {
+        theme: "night"
+      },
+      fallback_policy: {
+        mode: "reuse"
+      },
+      version: 3
+    })
+
+    const cloned = clonePresentationVisualStyleSnapshot(original)
+
+    expect(cloned).toEqual(original)
+    expect(cloned).not.toBe(original)
+    expect(cloned?.tags).not.toBe(original.tags)
+    expect(cloned?.generation_rules).not.toBe(original.generation_rules)
+    expect(cloned?.appearance_defaults).not.toBe(original.appearance_defaults)
+    expect(cloned?.fallback_policy).not.toBe(original.fallback_policy)
+  })
+
+  it("normalizes persona profile and exemplar ids, buddy summary, and tags", () => {
+    const profile = normalizePersonaProfile({
+      persona_id: 42,
+      buddySummary: {
+        personaName: " Garden Helper ",
+        hasBuddy: "0",
+        roleSummary: " research companion ",
+        visual: {
+          speciesId: "fox",
+          silhouetteId: "slim",
+          paletteId: "moss"
+        }
+      }
+    })
+
+    const exemplar = normalizePersonaExemplar({
+      id: 9,
+      personaId: "42",
+      kind: "voice",
+      scenarioTags: [" gardening ", "", "planning"],
+      capabilityTags: [" research ", " "],
+      priority: "4",
+      notes: " keep this trimmed ",
+      created_at: "2026-04-17T00:00:00Z",
+      last_modified: "2026-04-17T01:00:00Z"
+    })
+
+    expect(profile).toMatchObject({
+      id: "42",
+      buddy_summary: {
+        has_buddy: false,
+        persona_name: "Garden Helper",
+        role_summary: "research companion",
+        visual: {
+          species_id: "fox",
+          silhouette_id: "slim",
+          palette_id: "moss"
+        }
+      }
+    })
+
+    expect(exemplar).toMatchObject({
+      id: "9",
+      persona_id: "42",
+      kind: "voice",
+      scenario_tags: ["gardening", "planning"],
+      capability_tags: ["research"],
+      priority: 4,
+      notes: " keep this trimmed ",
+      created_at: "2026-04-17T00:00:00Z",
+      last_modified: "2026-04-17T01:00:00Z"
+    })
+  })
+})
